Extracted auctionCreated embed.
