+++ events/customEvents/nouns/auctionCreated.js
@@ -2,6 +2,9 @@ const { MessageEmbed, Channel } = require('discord.js');
 const { hyperlink } = require('@discordjs/builders');
 
 const Logger = require('../../../helpers/logger');
+const {
+   generateAuctionCreatedEmbed,
+} = require('../../../views/embeds/auctionCreated');
 
 module.exports = {
    name: 'auctionCreated',
@@ -10,52 +13,34 @@ module.exports = {
     * @param {Channel} genChannel
     */
    async execute(genChannel, data) {
+      Logger.info(
+         'events/nouns/auctionCreated.js: Attempting to handle an auction creation event.',
+         {
+            id: `${data.id}`,
+            startTime: data.startTime,
+            endTime: data.endTime,
+         },
+      );
+
       try {
-         Logger.info(
-            'events/nouns/auctionCreated.js: Attempting to handle an auction creation event.',
-            {
-               id: `${data.id}`,
-               startTime: data.startTime,
-               endTime: data.endTime,
-            },
-         );
-
-         const { id, startTime, endTime } = data;
-
-         const nounsWTF = hyperlink(
-            'Nouns.wtf',
-            `https://nouns.wtf/noun/${id}`,
-         );
-         const pronouns = hyperlink(
-            'Pronouns.gg',
-            `https://pronouns.gg/noun/${id}`,
-         );
-         const nounOClock = hyperlink(
-            'Nounoclock.app',
-            `https://www.nounoclock.app/`,
-         );
-
-         const acEmbed = new MessageEmbed()
-            .setColor('#00FFFF')
-            .setTitle(`New Auction | Noun ${id}`)
-            .setDescription(`${nounsWTF}\n${pronouns}\n${nounOClock}`)
-
-            .setImage(`https://noun.pics/${id}.png`);
-
-         Logger.info(
-            'events/nouns/auctionCreated.js: Successfully handled an auction creation event.',
-            {
-               id: data.id,
-               startTime: data.startTime,
-               endTime: data.endTime,
-            },
-         );
-
-         return await genChannel.send({ embeds: [acEmbed] });
+         const embed = generateAuctionCreatedEmbed(data);
+         await genChannel.send({ embeds: [embed] });
       } catch (error) {
-         Logger.info('events/nouns/auctionCreated.js: Received an error.', {
-            error: error,
-         });
+         return Logger.error(
+            'events/customEvents/nouns/auctionCreated.js: Received an error.',
+            {
+               error: error,
+            },
+         );
       }
+
+      Logger.info(
+         'events/nouns/auctionCreated.js: Successfully handled an auction creation event.',
+         {
+            id: data.id,
+            startTime: data.startTime,
+            endTime: data.endTime,
+         },
+      );
    },
 };
