+++ views/embeds/auctionCreated.js
@@ -1,0 +1,22 @@
+const { MessageEmbed } = require('discord.js');
+const { hyperlink } = require('@discordjs/builders');
+
+exports.generateAuctionCreatedEmbed = function (data) {
+   const nounsWTF = hyperlink('Nouns.wtf', `https://nouns.wtf/noun/${data.id}`);
+   const pronouns = hyperlink(
+      'Pronouns.gg',
+      `https://pronouns.gg/noun/${data.id}`,
+   );
+   const nounOClock = hyperlink(
+      'Nounoclock.app',
+      `https://www.nounoclock.app/`,
+   );
+
+   const embed = new MessageEmbed()
+      .setColor('#00FFFF')
+      .setTitle(`New Auction | Noun ${id}`)
+      .setDescription(`${nounsWTF}\n${pronouns}\n${nounOClock}`)
+      .setImage(`https://noun.pics/${id}.png`);
+
+   return embed;
+};
